Move pending transaction insertion out of state index
