+++ src/raps/actions/unlock.ts
@@ -11,7 +11,7 @@ import { type ChainId } from '@/features/network/types/backendNetworks';
 import { getProvider, toHex } from '@/handlers/web3';
 import { ensureError, logger, RainbowError } from '@/logger';
 import { ETH_ADDRESS } from '@/references/constants';
-import { addNewTransaction } from '@/state/pendingTransactions';
+import { addNewTransaction } from '@/state/pendingTransactions/addNewTransaction';
 import { type Call } from '@rainbow-me/sdk';
 
 import { type ActionProps, type PrepareActionProps, type RapActionResult, type RapUnlockActionParameters } from '../references';
